Create separate component for static search page.
